crunchtime Edit #5
Database stablized:

oddities: if you add one of the playlist video into favorites, app
crashes and follow up opening the database also causes a crash.

on emulator pressing on the video let's you go to the video but on
phone, it can't detect for some reason.

Menu's all same across. Can't apply it to EmbedActivity for some reason